Review corrections
