+++ src/attributes/components/AttributeSwatchField/messages.ts
@@ -1,0 +1,12 @@
+import { defineMessages } from "react-intl";
+
+export const swatchFieldMessages = defineMessages({
+  image: {
+    defaultMessage: "Image",
+    description: "swatch attribute image label"
+  },
+  picker: {
+    defaultMessage: "Picker",
+    description: "swatch attribute color picker label"
+  }
+});
