+++ src/components/Attributes/messages.ts
@@ -1,0 +1,16 @@
+import { defineMessages } from "react-intl";
+
+export const attributeRowMessages = defineMessages({
+  multipleValueLabel: {
+    defaultMessage: "Values",
+    description: "attribute values"
+  },
+  valueLabel: {
+    defaultMessage: "Value",
+    description: "attribute value"
+  },
+  reference: {
+    defaultMessage: "Assign references",
+    description: "button label"
+  }
+});
