+++ src/components/Attributes/types.ts
@@ -1,5 +1,34 @@
+import { AttributeValueFragment } from "@saleor/fragments/types/AttributeValueFragment";
+import { PageErrorWithAttributesFragment } from "@saleor/fragments/types/PageErrorWithAttributesFragment";
+import { ProductErrorWithAttributesFragment } from "@saleor/fragments/types/ProductErrorWithAttributesFragment";
+import { FormsetChange } from "@saleor/hooks/useFormset";
+import { FetchMoreProps, ReorderEvent } from "@saleor/types";
+
+import { AttributeInput } from "./Attributes";
+
 export enum VariantAttributeScope {
   ALL = "ALL",
   VARIANT_SELECTION = "VARIANT_SELECTION",
   NOT_VARIANT_SELECTION = "NOT_VARIANT_SELECTION"
+}
+
+export interface AttributeRowHandlers {
+  onChange: FormsetChange<string>;
+  onFileChange: FormsetChange<File>;
+  onMultiChange: FormsetChange<string>;
+  onReferencesAddClick: (attribute: AttributeInput) => void;
+  onReferencesRemove: FormsetChange<string[]>;
+  onReferencesReorder: FormsetChange<ReorderEvent>;
+  fetchAttributeValues: (query: string, attributeId: string) => void;
+  fetchMoreAttributeValues: FetchMoreProps;
+}
+
+export interface AttributeRowProps extends AttributeRowHandlers {
+  attribute: AttributeInput;
+  attributeValues: AttributeValueFragment[];
+  disabled: boolean;
+  error: ProductErrorWithAttributesFragment | PageErrorWithAttributesFragment;
+  loading: boolean;
+  entityId: string;
+  onAttributeSelectBlur?: () => void;
 }
